# 27-fold speedup for Pol.Deg