--- a/go.mod
+++ b/go.mod
@@ -1,1 +1,3 @@
 module github.com/restic/chunker
+
+go 1.9